Extract makeplan into detail/makeplan.hpp and rename C++ .h → .hpp
- Rename internal C++ headers to .hpp: finufft_core.h, spreadinterp.h,
  fft.h, test_defs.h (public C headers keep .h)
- Replace #ifndef/#define include guards with #pragma once
- Extract constructor, finufft_default_opts_t, finufft_makeplan_t,
  setup_spreadinterp, precompute_horner_coeffs, init_grid_kerFT_FFT,
  and helper math (set_nf_type12, onedim_fseries_kernel) from
  finufft_core.cpp into include/finufft/detail/makeplan.hpp
- Add src/makeplan.cpp with explicit template instantiations
- Default the destructor (~FINUFFT_PLAN_T = default)
- Update all #include directives across the codebase
- Update src/CMakeLists.txt and makefile

This enables parallel compilation of makeplan.cpp and finufft_core.cpp,
and establishes the detail/ header pattern for future splits (setpts,
execute, spreadinterp).
