+++ devel/CMakeLists.txt
@@ -25,5 +25,3 @@ add_executable(padding padding.cpp)
 target_compile_features(padding PRIVATE cxx_std_17)
 target_link_libraries(padding finufft xsimd)
 target_compile_options(padding PRIVATE -march=native)
-add_executable(guru_benchmark guru_benchmark.cpp)
-target_link_libraries(guru_benchmark finufft benchmark)
